Reorganize directory
The idea is to mirror the standard Cordova structure
