+++ src/cca/www/config.xml
@@ -1,7 +1,7 @@
 <?xml version='1.0' encoding='utf-8'?>
 <widget id="org.uproxy" version="0.0" xmlns="http://www.w3.org/ns/widgets" xmlns:cdv="http://cordova.apache.org/ns/1.0" xmlns:android="http://schemas.android.com/apk/res/android">
-    <preference name="SplashScreen" value="splashscreen" />
-    <preference name="SplashScreenDelay" value="20000" />
+    <preference name="SplashScreen" value="www/splashscreen" />
+    <preference name="SplashScreenDelay" value="2000" />
     <preference name="SplashMaintainAspectRatio" value="true" />
     <platform name="ios">
         <preference name="deployment-target" value="7.0"/>
