+++ src/cca/www/manifest.json
