+++ src/cca/www/model/vpn_device.ts
