+++ src/cca/www/scripts/cordova_browser_api.ts
