+++ src/cca/www/scripts/cordova_core_connector.ts
