+++ src/cca/www/scripts/iosrtc-register-globals.js
