+++ src/cca/www/scripts/main.core-env.ts
